double ftw

+++ src/character.h
@@ -44,7 +44,7 @@ public:
 	double getDefValue();
 	int wsc() { return wearing.weapon.getWeaponStaminaConsumption(); };
 	int ssc() { return wearing.shield.getShieldStaminaConsumption(); };
-	int gdx() { return wearing.weapon.getDamageX(); };
+	double gdx() { return wearing.weapon.getDamageX(); };
 	void BrokenInv();
 };
